update to remove classes inside unitTests
